Code modified in order for release version to work correctly
The code is now tested so the user can run the program in batch including .xls file, mongo host and .csv file.
To do this database.properties had to not be used as when executed from cmd it could not locate the file. We should correct this problem in the future.